Rename queryrange to querymiddleware
