+++ pkg/frontend/querymiddleware/model.proto
@@ -5,9 +5,10 @@
 
 syntax = "proto3";
 
+// Uses queryrange package to offer backwards compatibility with cached extents.
 package queryrange;
 
-option go_package = "queryrange";
+option go_package = "querymiddleware";
 
 import "github.com/gogo/protobuf/gogoproto/gogo.proto";
 import "google/protobuf/duration.proto";
