+++ pkg/frontend/querymiddleware/roundtrip.go
@@ -3,7 +3,7 @@
 // Provenance-includes-license: Apache-2.0
 // Provenance-includes-copyright: The Prometheus Authors.
 
-package queryrange
+package querymiddleware
 
 import (
 	"context"
