+++ pkg/frontend/querymiddleware/stats.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package queryrange
+package querymiddleware
 
 import (
 	"context"
