+++ pkg/frontend/querymiddleware/step_align_test.go
@@ -3,7 +3,7 @@
 // Provenance-includes-license: Apache-2.0
 // Provenance-includes-copyright: The Cortex Authors.
 
-package queryrange
+package querymiddleware
 
 import (
 	"context"
